move MultiStorage and replace mock
