+++ lib/MultiStorage/__mocks__/index.js
@@ -1,6 +1,6 @@
-import MemoryOnlyProvider, {mockStore, mockSet, setMockStore} from '../providers/MemoryOnlyProvider';
+import MemoryOnlyProvider, {mockStore, mockSet, setMockStore} from '../../storage/providers/MemoryOnlyProvider';
 
-const provider = {
+const MultiStorageMock = {
     init: jest.fn(MemoryOnlyProvider.init),
     getItem: jest.fn(MemoryOnlyProvider.getItem),
     multiGet: jest.fn(MemoryOnlyProvider.multiGet),
@@ -19,4 +19,4 @@ const provider = {
     setMockStore: jest.fn((data) => setMockStore(data)),
 };
 
-export default provider;
+export default MultiStorageMock;
