+++ lib/MultiStorage/index.js
@@ -1,24 +1,19 @@
 import _ from 'underscore';
-import PlatformStorage from './storage';
-import MemoryOnlyProvider from './storage/providers/MemoryOnlyProvider';
+import PlatformStorage from '../storage';
+import MemoryOnlyProvider from '../storage/providers/MemoryOnlyProvider';
 
 let storageProviders = [MemoryOnlyProvider, PlatformStorage];
 
 function runConsecutively(operationName, params) {
-    let value;
-    const providerCount = storageProviders.length;
-    const provider = storageProviders[0];
+    const runOperation = (providerIndex) => {
+        const provider = storageProviders[providerIndex];
 
-    const runOperation = (providerIndex) =>
-        new Promise((resolve, reject) => {
+        return new Promise((resolve, reject) => {
             provider[operationName](...params)
-                .then((val) => {
-                    value = val;
-                    resolve(value);
-                })
-                .catch(() => {
-                    if (providerIndex + 1 >= providerCount) {
-                        reject();
+                .then(resolve)
+                .catch((error) => {
+                    if (providerIndex + 1 >= storageProviders.length) {
+                        reject(error);
                         return;
                     }
 
@@ -27,6 +22,7 @@ function runConsecutively(operationName, params) {
                         .catch(reject);
                 });
         });
+    };
 
     return runOperation(0);
 }
@@ -34,7 +30,7 @@ function runConsecutively(operationName, params) {
 function runInParallel(operationName, resolveAll = false, ...params) {
     const promises = _.map(storageProviders, (provider) => provider[operationName](...params));
 
-    return resolveAll ? promises[0] : Promise.all(promises).then((values) => values[0]);
+    return resolveAll ? Promise.all(promises).then((values) => values[0]) : promises[0];
 }
 
 const MultiStorage = {
